+++ NEWS.rst
@@ -4,7 +4,10 @@ NEWS for system-image updater
 
 0.3 (2013-XX-XX)
 ================
-
+ * Update the client.ini file to reflect the actual update service (which is
+   now deployed) and the system partitioning on the actual device.
+ * By default, search for client.ini in /etc/system-image/client.ini
+   (LP: #1199177)
 
 0.2 (2013-06-27)
 ================